+++ js/components/generate/__tests__/generate.test.js
@@ -10,21 +10,19 @@ describe("Handle invalid course inputs appropriately", () => {
   it.each([
     {
       coursesInputText: "MAT777H1, asdasdasd, CSC876Y1",
-      expectedWarning: "The course asdasdasd was invalid! Please check your input.",
+      expectedWarning: "Invalid course code: asdasdasd",
     },
     {
       coursesInputText: "MAT777H1",
-      expectedWarning: "The course MAT777H1 was invalid! Please check your input.",
+      expectedWarning: "Invalid course code: MAT777H1",
     },
     {
       coursesInputText: "CSC110Y1, SDS777H1, CSC343H1, MAT888Y1, MAT237Y1",
-      expectedWarning:
-        "The courses [SDS777H1, MAT888Y1] were invalid! Please check your input.",
+      expectedWarning: "Invalid course codes: SDS777H1, MAT888Y1",
     },
     {
       coursesInputText: "CSC110Y1 CSC343H1",
-      expectedWarning:
-        "The course CSC110Y1 CSC343H1 was invalid! Please check your input.",
+      expectedWarning: "Invalid course code: CSC110Y1 CSC343H1",
     },
     {
       coursesInputText: "",
@@ -45,7 +43,7 @@ describe("Handle invalid course inputs appropriately", () => {
     await user.keyboard(coursesInputText)
 
     expect(screen.queryByText(expectedWarning)).toBeNull()
-    const genButton = screen.getByText("Generate Graph")
+    const genButton = screen.getByText("Generate")
     await user.click(genButton)
 
     const errorMessage = await screen.findByText(expectedWarning)
@@ -61,16 +59,15 @@ describe("Handle invalid department inputs appropriately", () => {
   it.each([
     {
       departmentInputText: "CSC, SDS, MAT, PHYS, BIO",
-      expectedWarning: "The department PHYS was invalid! Please check your input.",
+      expectedWarning: "Invalid department: PHYS",
     },
     {
       departmentInputText: "csc, abcd, SDS, MAT, PHYS, BIO",
-      expectedWarning:
-        "The departments [abcd, PHYS] were invalid! Please check your input.",
+      expectedWarning: "Invalid departments: abcd, PHYS",
     },
     {
       departmentInputText: "CSC, MAT STA",
-      expectedWarning: "The department MAT STA was invalid! Please check your input.",
+      expectedWarning: "Invalid department: MAT STA",
     },
   ])(".$departmentInputText", async ({ departmentInputText, expectedWarning }) => {
     const user = userEvent.setup()
@@ -89,7 +86,7 @@ describe("Handle invalid department inputs appropriately", () => {
     await user.keyboard(coursesInputText)
 
     expect(screen.queryByText(expectedWarning)).toBeNull()
-    const genButton = screen.getByText("Generate Graph")
+    const genButton = screen.getByText("Generate")
     await user.click(genButton)
 
     const errorMessage = await screen.findByText(expectedWarning)
@@ -105,22 +102,20 @@ describe("Handle invalid taken courses inputs appropriately", () => {
   it.each([
     {
       takenCoursesInputText: "MAT777H1, asdasdasd, CSC876Y1",
-      expectedWarning: "The course asdasdasd was invalid! Please check your input.",
+      expectedWarning: "Invalid course code: asdasdasd",
     },
     {
       takenCoursesInputText: "CSC110Y1 CSC343H1",
-      expectedWarning:
-        "The course CSC110Y1 CSC343H1 was invalid! Please check your input.",
+      expectedWarning: "Invalid course code: CSC110Y1 CSC343H1",
     },
     {
       takenCoursesInputText: "MAT1234H1, CSC207H1, CSC1234Y1",
-      expectedWarning:
-        "The courses [MAT1234H1, CSC1234Y1] were invalid! Please check your input.",
+      expectedWarning: "Invalid course codes: MAT1234H1, CSC1234Y1",
     },
   ])(".$takenCoursesInputText", async ({ takenCoursesInputText, expectedWarning }) => {
     const user = userEvent.setup()
     const takenCoursesInputField = screen.getByPlaceholderText(
-      "E.g., CSC207H1, CSC236H1"
+      "e.g., CSC207H1, CSC236H1"
     )
     await user.click(takenCoursesInputField)
     await user.tripleClick(takenCoursesInputField)
@@ -136,7 +131,7 @@ describe("Handle invalid taken courses inputs appropriately", () => {
     await user.keyboard(coursesInputText)
 
     expect(screen.queryByText(expectedWarning)).toBeNull()
-    const genButton = screen.getByText("Generate Graph")
+    const genButton = screen.getByText("Generate")
     await user.click(genButton)
 
     const errorMessage = await screen.findByText(expectedWarning)
@@ -153,7 +148,7 @@ it("No warning for valid course input strings", async () => {
   await user.tripleClick(coursesInputField)
   await user.keyboard(coursesInputText)
   expect(screen.queryByText("Invalid Course Input")).toBeNull()
-  const genButton = screen.getByText("Generate Graph")
+  const genButton = screen.getByText("Generate")
   await user.click(genButton)
   await expect(screen.findByText(/invalid/i)).rejects.toThrow()
 })
@@ -167,7 +162,7 @@ it("Submitting with valid courses and then making them invalid correctly updates
   await user.tripleClick(coursesInputField)
   await user.keyboard(coursesInputText)
   expect(screen.queryByText("Invalid Course Input")).toBeNull()
-  let genButton = screen.getByText("Generate Graph")
+  let genButton = screen.getByText("Generate")
   await user.click(genButton)
   await expect(screen.findByText("Invalid Course Input")).rejects.toThrow()
   expect(screen.getByText("CSC443H1")).toBeDefined()
@@ -179,7 +174,7 @@ it("Submitting with valid courses and then making them invalid correctly updates
   await user.keyboard(coursesInputTextBad)
   expect(screen.queryByText(/invalid/i)).toBeNull()
   await user.click(genButton)
-  const errorMessage = await screen.findByText(/The course CSC443H7 was invalid/i)
+  const errorMessage = await screen.findByText(/Invalid course code: CSC443H7/i)
   expect(errorMessage).toBeDefined()
   expect(screen.queryByText("CSC443H1")).toBeNull()
 })